rename: 카드 등록 파일명 수정
