+++ src/pages/CardRegistration/CardRegistration.style.ts
